+++ src/platforms/joomla/Framework/Theme.php
@@ -111,6 +111,12 @@ class Theme extends AbstractTheme
             if ($filename) {
                 $lang->load("tpl_{$this->name}_positions", dirname(dirname(dirname($filename))), 'en-GB');
             }
+
+            // Load template language files, including overrides.
+            $paths = $locator->findResources('gantry-theme://language');
+            foreach (array_reverse($paths) as $path) {
+                $lang->load("tpl_{$this->name}", dirname($path));
+            }
         }
 
         $doc = \JFactory::getDocument();